+++ src/corelib/plugin/qlibrary.cpp
@@ -270,7 +270,7 @@ static bool findPatternUnloaded(const QString &library, QLibraryPrivate *lib)
     qsizetype pos = 0;
     char pattern[] = "qTMETADATA ";
     pattern[0] = 'Q'; // Ensure the pattern "QTMETADATA" is not found in this library should QPluginLoader ever encounter it.
-    const ulong plen = qstrlen(pattern);
+    const ulong plen = ulong(qstrlen(pattern));
 #if defined (Q_OF_ELF) && defined(Q_CC_GNU)
     int r = QElfParser().parse(filedata, fdlen, library, lib, &pos, &fdlen);
     if (r == QElfParser::Corrupt || r == QElfParser::NotElf) {
